fixed transaction tables to be generic transaction tables
